Fix visualization link to impact-analysis.tsx
Update the file paths and homepage for the visualization dashboard.

* Rename `visualizations-backup/src/components/impact-analysis.tsx` to `visualizations/src/components/impact-analysis.tsx`.
* Update the `homepage` field in `visualizations/package.json` to `"https://tim-spurlin.github.io/churches-united-sustainability-initiative"`.
* Add `"predeploy": "npm run build"` and `"deploy": "gh-pages -d build"` to the `scripts` section in `visualizations/package.json`.
* Update the link to `impact-analysis.tsx` in `README.md` to point to the correct path.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/Tim-Spurlin/churches-united-sustainability-initiative?shareId=XXXX-XXXX-XXXX-XXXX).
