+++ visualizations/package.json
@@ -53,5 +53,5 @@
   "devDependencies": {
     "gh-pages": "^6.3.0"
   },
-  "homepage": "https://tim-spurlin.github.io/churches-united-sustainability-initiative/visualizations"
+  "homepage": "https://tim-spurlin.github.io/churches-united-sustainability-initiative"
 }
